refactor: 패키지 구조 정리
- setup 패키지 생성
