+++ src/main/java/janggi/controller/JanggiController.java
@@ -1,9 +1,9 @@
 package janggi.controller;
 
-import janggi.domain.SetupPolicy;
 import janggi.domain.board.Board;
 import janggi.domain.board.BoardGenerator;
 import janggi.domain.command.SetupCommand;
+import janggi.domain.setup.SetupPolicy;
 import janggi.domain.team.BlueTeam;
 import janggi.domain.team.RedTeam;
 import janggi.domain.team.Team;
