+++ src/test/java/janggi/domain/BoardGeneratorTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import janggi.domain.board.Board;
 import janggi.domain.board.BoardGenerator;
+import janggi.domain.setup.InnerElephantSetupPolicy;
 import janggi.domain.team.BlueTeam;
 import janggi.domain.team.RedTeam;
 import org.assertj.core.api.InstanceOfAssertFactories;
