+++ src/test/java/janggi/domain/CannonTest.java
@@ -42,22 +42,22 @@ class CannonTest {
         @DisplayName("포는 기물을 뛰어 넘어서 이동할 수 있다.")
         void success_1() {
             positionPieceMap = Map.of(
-                Position.valueOf(6, 7), cannon,
-                Position.valueOf(6, 5), enemySoldier,
-                Position.valueOf(7, 7), allySoldier);
+                    Position.valueOf(6, 7), cannon,
+                    Position.valueOf(6, 5), enemySoldier,
+                    Position.valueOf(7, 7), allySoldier);
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
             List<Position> expected = List.of(
-                Position.valueOf(6, 1),
-                Position.valueOf(6, 2),
-                Position.valueOf(6, 3),
-                Position.valueOf(6, 4),
-                Position.valueOf(8, 7),
-                Position.valueOf(9, 7),
-                Position.valueOf(10, 7));
+                    Position.valueOf(6, 1),
+                    Position.valueOf(6, 2),
+                    Position.valueOf(6, 3),
+                    Position.valueOf(6, 4),
+                    Position.valueOf(8, 7),
+                    Position.valueOf(9, 7),
+                    Position.valueOf(10, 7));
 
             List<Position> actual = cannon.calculateMovablePositions(Position.valueOf(6, 7),
-                boardMediator);
+                    boardMediator);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
@@ -66,19 +66,19 @@ class CannonTest {
         @DisplayName("포는 포를 넘을 수 없다.")
         void success_2() {
             positionPieceMap = Map.of(
-                Position.valueOf(6, 7), cannon,
-                Position.valueOf(4, 7), enemySoldier,
-                Position.valueOf(6, 5), enemyCannon,
-                Position.valueOf(9, 7), allyCannon);
+                    Position.valueOf(6, 7), cannon,
+                    Position.valueOf(4, 7), enemySoldier,
+                    Position.valueOf(6, 5), enemyCannon,
+                    Position.valueOf(9, 7), allyCannon);
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
             List<Position> expected = List.of(
-                Position.valueOf(1, 7),
-                Position.valueOf(2, 7),
-                Position.valueOf(3, 7));
+                    Position.valueOf(1, 7),
+                    Position.valueOf(2, 7),
+                    Position.valueOf(3, 7));
 
             List<Position> actual = cannon.calculateMovablePositions(Position.valueOf(6, 7),
-                boardMediator);
+                    boardMediator);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
@@ -87,15 +87,15 @@ class CannonTest {
         @DisplayName("포는 포를 잡을 수 없다.")
         void success_3() {
             positionPieceMap = Map.of(
-                Position.valueOf(6, 7), cannon,
-                Position.valueOf(6, 3), enemyCannon,
-                Position.valueOf(6, 4), allySoldier);
+                    Position.valueOf(6, 7), cannon,
+                    Position.valueOf(6, 3), enemyCannon,
+                    Position.valueOf(6, 4), allySoldier);
             Board board = new Board(positionPieceMap);
             BoardMediator boardMediator = new BoardMediatorImpl(board);
             List<Position> expected = List.of();
 
             List<Position> actual = cannon.calculateMovablePositions(Position.valueOf(6, 7),
-                boardMediator);
+                    boardMediator);
 
             assertThat(actual).hasSameElementsAs(expected);
         }
